// src/lib/api/generated/models/getAdminPromotionalBannersParams.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */
import type { GetAdminPromotionalBannersStatus } from './getAdminPromotionalBannersStatus';

export type GetAdminPromotionalBannersParams = {
page?: number;
per_page?: number;
position?: string;
status?: GetAdminPromotionalBannersStatus;
};