// src/lib/api/generated/models/getAdminPromotionalBannersStatus.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * ZidoBid Admin API
 * OpenAPI spec version: 0.1.0-local
 */

export type GetAdminPromotionalBannersStatus = typeof GetAdminPromotionalBannersStatus[keyof typeof GetAdminPromotionalBannersStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetAdminPromotionalBannersStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;